requirements.txt
beautifulsoup4==4.9.3
click==8.0.1
feedwerk==1.0.0
Flask==2.0.1
Flask-Ext==0.1
Flask-Gravatar==0.5.0
Flask-WTF==0.15.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML
six==1.16.0
soupsieve==2.2.1
Werkzeug==2.0.1
WTForms==2.3.3
